# docs/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "job-tracker-api",
    "description": "REST API for job-tracker-dashboard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/loops/{loop_id}/matches/import-preview": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Preview one vacancy URL for a loop without saving it",
        "operationId": "import_match_preview_api_v1_loops__loop_id__matches_import_preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyImportPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyImportPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Save vacancy preview as a match under a loop",
        "operationId": "create_match_api_v1_loops__loop_id__matches_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyMatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyMatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "List vacancy matches for a loop",
        "operationId": "list_matches_api_v1_loops__loop_id__matches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "saved",
                    "ignored",
                    "converted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyMatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/from-preview": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Save one selected discovery preview item as a vacancy match",
        "operationId": "create_match_from_preview_api_v1_loops__loop_id__matches_from_preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyMatchFromPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyMatchFromPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/preview-ignores": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Hide one discovery preview item from future preview feeds",
        "operationId": "create_preview_ignore_api_v1_loops__loop_id__matches_preview_ignores_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyPreviewIgnoreRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyPreviewIgnoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "List preview items hidden by the current user for a loop",
        "operationId": "list_preview_ignores_api_v1_loops__loop_id__matches_preview_ignores_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyPreviewIgnoreListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/preview-ignores/{ignore_id}": {
      "delete": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Remove one hidden discovery preview item",
        "operationId": "delete_preview_ignore_api_v1_loops__loop_id__matches_preview_ignores__ignore_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "ignore_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ignore Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/{match_id}": {
      "patch": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Update editable vacancy match fields",
        "operationId": "patch_match_api_v1_loops__loop_id__matches__match_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyMatchPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyMatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/{match_id}/evaluate": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Evaluate deterministic score and duplicate metadata for a vacancy match",
        "operationId": "evaluate_match_api_v1_loops__loop_id__matches__match_id__evaluate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyMatchEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/{match_id}/create-application": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Create an application from a saved vacancy match",
        "operationId": "create_application_from_match_api_v1_loops__loop_id__matches__match_id__create_application_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationFromMatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApplicationFromMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/{match_id}/convert-to-application": {
      "post": {
        "tags": [
          "vacancy-matches"
        ],
        "summary": "Create an application from a saved vacancy match",
        "operationId": "convert_match_to_application_api_v1_loops__loop_id__matches__match_id__convert_to_application_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VacancyImportPreviewRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "VacancyImportPreviewRequest"
      },
      "VacancyImportPreviewResponse": {
        "properties": {
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "vacancy_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vacancy Description"
          },
          "confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_url",
          "source",
          "company_name",
          "role_title",
          "location_text",
          "vacancy_description",
          "confidence",
          "warnings"
        ],
        "title": "VacancyImportPreviewResponse"
      },
      "VacancyMatchCreate": {
        "properties": {
          "source_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Source Url"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "vacancy_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vacancy Description"
          },
          "raw_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Metadata"
          },
          "confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "saved",
              "ignored",
              "converted"
            ],
            "title": "Status",
            "default": "saved"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_url"
        ],
        "title": "VacancyMatchCreate"
      },
...
      "VacancyMatchRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "loop_id": {
            "type": "string",
            "title": "Loop Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "vacancy_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vacancy Description"
          },
          "raw_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Metadata"
          },
          "confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "application_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "loop_id",
          "source_url",
          "source",
          "company_name",
          "role_title",
          "location_text",
          "vacancy_description",
          "confidence",
          "warnings",
          "status",
          "application_id",
          "created_at",
          "updated_at"
        ],
        "title": "VacancyMatchRead"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}